# Refactor the `Validation::IsValidUTF8()` method