--- a/binding.gyp
+++ b/binding.gyp
@@ -3,8 +3,6 @@
     {
       'target_name': 'validation',
       'include_dirs': ["<!(node -e \"require('nan')\")"],
-      'cflags!': [ '-O3' ],
-      'cflags': [ '-O2' ],
       'sources': [ 'src/validation.cc' ],
       'xcode_settings': {
         'MACOSX_DEPLOYMENT_TARGET': '10.8',