This patch simplifies the C++ code taking advantage of the [routine](https://www.cl.cam.ac.uk/%7Emgk25/ucs/utf8_check.c) written by Markus Kuhn.