@lpinca 2016-12-08
This also makes the validation ~2x faster:

Before

```
10000 roundtrips of 64 B text data:	1.5s	421.78 KiB/s
5000 roundtrips of 16 KiB text data:	2.9s	27.25 MiB/s
1000 roundtrips of 128 KiB text data:	3.9s	32.02 MiB/s
100 roundtrips of 1 MiB text data:	2.6s	38 MiB/s
```

After

```
10000 roundtrips of 64 B text data:	1.5s	412.68 KiB/s
5000 roundtrips of 16 KiB text data:	1.4s	56.69 MiB/s
1000 roundtrips of 128 KiB text data:	1.5s	81.32 MiB/s
100 roundtrips of 1 MiB text data:	1.5s	65.32 MiB/s
```