@lpinca 2016-12-08
As discussed in https://github.com/websockets/utf-8-validate/pull/32#issuecomment-258422921 this could be optimized further but for now I'm happy with it as is.

@websockets/admin if you can, please take a look. If there are no objections I will merge this and https://github.com/websockets/utf-8-validate/pull/33 and cut a new major version.

Technically there are no breaking changes, but I'm not sure if this catches all the cases that were previously handled so bumping the major version makes more sense.