@3rd-Eden 2016-12-08
@lpinca As long as it passes the autobahn tests with this validation change, i would see no objection.