@lpinca 2016-12-08
I'm pretty sure it does. I'll run the Autobahn test suite with this later to be 100% sure.