@lpinca 2016-12-08
All tests in the Autobahn test suite passed, merging.